serve: outsource libraries and tags
